More stuff

+++ survpi-master/uninstall.sh
@@ -1,3 +1,8 @@
+if [ "$PWD" != "/home/pi/SurveillancePi" ] then
+    echo "Please run the install script inside the /home/pi/SurveillancePi folder."
+    exit
+fi
+
 echo "Stopping service..."
 systemctl stop survpi
 systemctl disable survpi
